+++ reflex-basic-host.nix
@@ -1,9 +1,9 @@
 { mkDerivation, base, dependent-map, dependent-sum, lens, mtl
-, primitive, ref-tf, reflex, stdenv, stm
+, primitive, ref-tf, reflex, stdenv, stm, witherable
 }:
 mkDerivation {
   pname = "reflex-basic-host";
-  version = "0.1";
+  version = "0.2";
   src = ./.;
   isLibrary = true;
   isExecutable = true;
@@ -11,7 +11,9 @@ mkDerivation {
     base dependent-map dependent-sum lens mtl primitive ref-tf reflex
     stm
   ];
-  executableHaskellDepends = [ base mtl reflex ];
+  executableHaskellDepends = [
+    base lens mtl ref-tf reflex witherable
+  ];
   description = "A basic `reflex` host for backend work";
   license = stdenv.lib.licenses.bsd3;
 }