+++ ChangeLog.md
@@ -5,6 +5,7 @@
 * `Reflex.Host.Basic.basicHostWithQuit`: Swap the values the
   `BasicGuest` should return, to facilitate easier `fmap`-ping.
 * Do not fork a new thread when starting the host.
+* All hosts now run in separate reflex timelines.
 
 ## 0.1